modifié :         include/linked_list/crud/create.h 	modifié :         include/linked_list/crud/delete.h 	modifié :         include/linked_list/crud/read.h 	modifié :         include/linked_list/new.h 	modifié :         src/linked_list/crud/create.c 	modifié :         src/linked_list/crud/delete.c 	modifié :         src/linked_list/crud/read.c 	modifié :         src/linked_list/new.c 	modifié :         src/main.c

+++ include/linked_list/crud/create.h
@@ -1,7 +1,7 @@
 #ifndef CREATE_H
 #define CREATE_H
 
-    void create_node(Node *linked_list_first_element, int *value);
-    void create_node_recursion(Node *linked_list_first_element, int *value);
+    void create_node(Node *linked_list_first_element, float *value);
+    void create_node_recursion(Node *linked_list_first_element, float *value);
 
 #endif
