+++ include/linked_list/crud/delete.h
@@ -1,6 +1,6 @@
 #ifndef DELETE_H
 #define DELETE_H
 
-    void delete_node(Node **linked_list_first_element, int hash_node_to_delete);
+    void delete_node(Node **linked_list_first_element, float hash_node_to_delete);
 
 #endif
