+++ include/linked_list/crud/read.h
@@ -1,7 +1,7 @@
 #ifndef READ_H
 #define READ_H
 
-    void print_node(Node *linked_list_first_element, int hash_node_to_print);
-    void print_node_recursion(Node *linked_list_first_element, int hash_node_to_print);
+    void print_node(Node *linked_list_first_element, float hash_node_to_print);
+    void print_node_recursion(Node *linked_list_first_element, float hash_node_to_print);
 
 #endif
